refactor: split server page exports
