+++ app/routes/bookmarks/index.tsx
@@ -16,36 +16,19 @@ import type {
 } from "../../db/models";
 import { $R, ROUTE_DEF } from "../../misc/routes";
 import { trpc } from "../../trpc/client";
-import {
-  disableUrlQueryRevalidation,
-  useTypedUrlQuery,
-} from "../../utils/loader-utils";
-import { makeLoader } from "../../utils/loader-utils.server";
+import { useTypedUrlQuery } from "../../utils/loader-utils";
 import { cls } from "../../utils/misc";
 import type { PageHandle } from "../../utils/page-handle";
 import type { CaptionEntry } from "../../utils/types";
 import { YoutubePlayer, usePlayerLoader } from "../../utils/youtube";
 import { CaptionEntryComponent, findCurrentEntry } from "../videos/$id";
 
+export * from "./index.server";
+
 export const handle: PageHandle = {
   navBarTitle: () => "Bookmarks",
   navBarMenu: () => <NavBarMenuComponent />,
 };
-
-//
-// loader
-//
-
-export const loader = makeLoader(async ({ ctx }) => {
-  await ctx.requireUser();
-  return null;
-});
-
-export const shouldRevalidate = disableUrlQueryRevalidation;
-
-//
-// component
-//
 
 export default function DefaultComponent() {
   const [urlQuery, setUrlQuery] = useTypedUrlQuery(
